% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HerbLeafAnalyzer.R
\name{calculateLeafArea}
\alias{calculateLeafArea}
\title{Calculate Leaf Area}
\usage{
calculateLeafArea(length, width)
}
\arguments{
\item{length}{Numeric, the length of the leaf in centimeters.}

\item{width}{Numeric, the width of the leaf in centimeters.}
}
\value{
Numeric, the calculated area of the leaf.
}
\description{
Calculates the area of a leaf based on provided length and width.
}
\examples{
calculateLeafArea(7, 3)
}
