% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HerbLeafAnalyzer.R
\name{displayHerbTable}
\alias{displayHerbTable}
\title{Display Herb Table}
\usage{
displayHerbTable(data)
}
\arguments{
\item{data}{A data frame containing herb data.}
}
\value{
The same data frame as provided in the input.
}
\description{
Returns the input data as a table.
}
\examples{
displayHerbTable(data)
}
